rag_library/requirements.txt
numpy
pandas
PyMuPDF
python-pptx
nltk
torch
transformers
sentence-transformers
faiss-cpu
chromadb
accelerate
